feat: add approval reducer
