+++ monitor/reducers/estateReducer.js
@@ -3,13 +3,11 @@ import { Log, env } from 'decentraland-commons'
 import { getParcelIdFromEvent, debouncedUpsertTileAsset } from './utils'
 import { Parcel, Estate } from '../../src/Asset'
 import { Publication } from '../../src/Listing'
-import { Approval } from '../../src/Approval'
 import { BlockTimestampService } from '../../src/BlockTimestamp'
 import { Tile } from '../../src/Tile'
 import { contractAddresses, eventNames } from '../../src/ethereum'
 import { ASSET_TYPES, decodeMetadata } from '../../shared/asset'
 import { getParcelMatcher, isEqualCoords } from '../../shared/parcel'
-import { isDuplicatedConstraintError } from '../../src/database'
 
 const log = new Log('estateReducer')
 const shouldUpdateCache = !env.get('SKIP_TILES_CACHE_UPDATE', false)
@@ -28,7 +26,7 @@ export async function estateReducer(event) {
 }
 
 async function reduceEstateRegistry(event) {
-  const { tx_hash, block_number, name, address } = event
+  const { tx_hash, block_number, name } = event
 
   switch (name) {
     case eventNames.CreateEstate: {
@@ -186,34 +184,6 @@ async function reduceEstateRegistry(event) {
         { operator: _approved.toLowerCase() },
         { id: estateId }
       )
-      break
-    }
-    case eventNames.ApprovalForAll: {
-      const _owner = event.args._owner.toLowerCase()
-      const _operator = event.args._operator.toLowerCase()
-      const _approved = event.args._approved === 'true'
-
-      try {
-        log.info(
-          `[${name}] ${_owner} ${
-            _approved ? 'set' : 'remove'
-          } ${_operator} as approved for all`
-        )
-        if (_approved) {
-          await Approval.approveForAll(address, _owner, _operator)
-        } else {
-          await Approval.delete({
-            token_address: address,
-            owner: _owner,
-            operator: _operator
-          })
-        }
-      } catch (error) {
-        if (!isDuplicatedConstraintError(error)) throw error
-        log.info(
-          `[${name}] ${_owner} has already set ${_operator} as approved for all`
-        )
-      }
       break
     }
     default:
