+++ src/ethereum/index.js
@@ -113,7 +113,7 @@ export function loadContracts() {
       eventNames: ['Transfer']
     },
     ERC721Bid: {
-      address: env.get('ERC721_BID_CONTRACT_ADDRESS').toLowerCase(),
+      address: getEnvAddress('ERC721_BID_CONTRACT_ADDRESS'),
       eventNames: ['BidCreated', 'BidAccepted', 'BidCancelled']
     }
   })
